Finalize authorization for reviews
