+++ views/campgrounds/show.ejs
@@ -36,6 +36,7 @@
     </div>
   </div>
   <div class="col-6">
+    <% if (currentUser) { %>
     <h2>Leave a Review</h2>
     <form
       action="/campgrounds/<%=campground._id%>/reviews"
@@ -68,17 +69,22 @@
       </div>
       <button class="btn btn-success">Submit</button>
     </form>
-    <% for(let review of campground.reviews) { %>
+    <% } %> <% for(let review of campground.reviews) { %>
     <div class="card mb-3">
       <div class="card-body">
         <h5 class="card-title">Rating: <%= review.rating %></h5>
+        <h6 class="card-subtitle mb-2 text-muted">
+          By <%= review.author.username %>
+        </h6>
         <p class="card-text">Review: <%= review.body %></p>
+        <% if(currentUser && review.author.equals(currentUser._id)) {%>
         <form
           action="/campgrounds/<%=campground._id%>/reviews/<%=review._id%>?_method=DELETE"
           method="POST"
         >
           <button class="btn btn-sm btn-danger">Delete</button>
         </form>
+        <% } %>
       </div>
     </div>
     <% } %>
